[wasm interpreter] Compute stack frame size as 64-bit
DrumBrake tracked a function's interpreter stack frame size (and the
underlying slot-offset counter) in 32-bit fields. For a function with a
very large frame, the byte size could exceed 2^32 and wrap to a small
value, slipping under the kMaxStackSize (32 MB) check in
WasmBytecode::InitializeSlots().

Widen slot_offset_ and total_frame_size_in_bytes_ (and frame_size() /
the WasmBytecode ctor argument) to size_t so the frame size is always
reported truthfully. An oversized frame now fails the stack check and
traps as a stack overflow at invocation.

Bug: 530655571
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/8055276
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#108517}

+++ src/wasm/interpreter/wasm-interpreter.cc
@@ -7420,7 +7420,7 @@ void WasmEHDataGenerator::RecordPotentialExceptionThrowingInstruction(
 }
 
 WasmBytecode::WasmBytecode(int func_index, const uint8_t* code_data,
-                           size_t code_length, uint32_t stack_frame_size,
+                           size_t code_length, size_t stack_frame_size,
                            const FunctionSig* signature,
                            const CanonicalSig* canonical_signature,
                            const InterpreterCode* interpreter_code,
