+++ src/wasm/interpreter/wasm-interpreter.h
@@ -182,7 +182,7 @@ static_assert(sizeof(WasmRef) == kSystemPointerSize);
 class WasmBytecode {
  public:
   WasmBytecode(int func_index, const uint8_t* code_data, size_t code_length,
-               uint32_t stack_frame_size, const FunctionSig* signature,
+               size_t stack_frame_size, const FunctionSig* signature,
                const CanonicalSig* canonical_signature,
                const InterpreterCode* interpreter_code, size_t blocks_count,
                const uint8_t* const_slots_data, size_t const_slots_length,
@@ -228,7 +228,7 @@ class WasmBytecode {
     return ref_slots_count_ - ref_rets_count_ - ref_args_count_;
   }
 
-  inline uint32_t frame_size() { return total_frame_size_in_bytes_; }
+  inline size_t frame_size() { return total_frame_size_in_bytes_; }
 
   static inline uint32_t ArgsSizeInSlots(const FunctionSig* sig);
   static inline uint32_t RetsSizeInSlots(const FunctionSig* sig);
@@ -272,7 +272,7 @@ class WasmBytecode {
   uint32_t rets_slots_size_;
   uint32_t locals_count_;
   uint32_t locals_slots_size_;
-  uint32_t total_frame_size_in_bytes_;
+  size_t total_frame_size_in_bytes_;
   uint32_t ref_args_count_;
   uint32_t ref_rets_count_;
   uint32_t ref_locals_count_;
@@ -1601,7 +1601,9 @@ class WasmBytecodeGenerator {
     return rets_slots_size_ + args_slots_size_;
   }
 
-  inline uint32_t GetStackFrameSize() const { return slot_offset_; }
+  inline uint32_t GetStackFrameSize() const {
+    return static_cast<uint32_t>(slot_offset_);
+  }
 
   uint32_t CurrentCodePos() const {
     return static_cast<uint32_t>(code_.size());
@@ -1871,13 +1873,14 @@ class WasmBytecodeGenerator {
       return CreateWasmRefSlot(value_type);
     }
     uint32_t slot_index = static_cast<uint32_t>(slots_.size());
-    slots_.push_back({value_type, slot_offset_, 0});
+    slots_.push_back({value_type, static_cast<uint32_t>(slot_offset_), 0});
     slot_offset_ += sizeof(T) / kSlotSize;
     return slot_index;
   }
   inline uint32_t CreateWasmRefSlot(ValueType value_type) {
     uint32_t slot_index = static_cast<uint32_t>(slots_.size());
-    slots_.push_back({value_type, slot_offset_, ref_slots_count_});
+    slots_.push_back(
+        {value_type, static_cast<uint32_t>(slot_offset_), ref_slots_count_});
     slot_offset_ += sizeof(WasmRef) / kSlotSize;
     ref_slots_count_++;
     return slot_index;
@@ -2093,7 +2096,12 @@ class WasmBytecodeGenerator {
   absl::flat_hash_map<Simd128, uint32_t, Simd128Hash> s128_const_cache_;
 
   std::vector<Simd128> simd_immediates_;
-  uint32_t slot_offset_;  // TODO(paolosev@microsoft.com): manage holes
+  // 64-bit so the running frame size cannot wrap for a pathologically large
+  // frame; such a frame then traps as a stack overflow at invocation (see
+  // WasmBytecode::InitializeSlots).
+  //
+  // TODO(paolosev@microsoft.com): manage holes
+  size_t slot_offset_;
 
   class RollbackStack {
    public:
